requirements.txt
fastapi
uvicorn
aiohttp
python-dotenv
python-telegram-bot==22.0
transformers
peft
nest_asyncio
beautifulsoup4
duckduckgo-search
datasets
pydantic
accelerate
scikit-learn
sqlite-utils
requests
pymupdf
python-docx
gradio
